requirements.txt
opencv-python>=4.5.0
numpy>=1.19.0
tensorflow>=2.5.0
pillow>=8.0.0